Move AQI logic from MainActivity to AQIFragment
